tycagent/requirements.txt
certifi==2020.12.5
chardet==4.0.0
debtcollector==2.2.0
idna==2.10
iso8601==0.1.14
keystoneauth1==4.3.1
msgpack==1.0.2
netaddr==0.8.0
netifaces==0.10.9
os-service-types==1.7.0
oslo.config==8.5.0
oslo.i18n==5.0.1
oslo.serialization==4.1.0
oslo.utils==4.8.0
packaging==20.9
pbr==5.5.1
pyparsing==2.4.7
python-keystoneclient==4.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rfc3986==1.4.0
six==1.15.0
stevedore==3.3.0
urllib3==1.26.4
wrapt==1.12.1
